+++ server/src/api/models/shape/shape.py
@@ -44,3 +44,7 @@ class ApiCoreShape(TypeIdModel):
     character: int | None = Field(..., typeId="CharacterId", noneAsNull=True)
     odd_hex_orientation: bool
     size: int
+    show_cells: bool
+    cell_fill_colour: str | None = Field(..., noneAsNull=True)
+    cell_stroke_colour: str | None = Field(..., noneAsNull=True)
+    cell_stroke_width: int | None = Field(..., noneAsNull=True)